+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unereleased]
 
+### Added
+
+- RouteSystem adapter
+
 ### Changed
 
 - refactored DataSource adapter for a new system API